cp: move providerName funcs to cputil
