+++ providers/component-protocol/utils/cputil/provider_name_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package base
+package cputil
 
 import (
 	"testing"
